fixed lib8tion moved library to own folder, pulled out the function definitions from the headers
