+++ src/lib8tion/ease8.cpp
@@ -1,5 +1,5 @@
-#include "qt/lib8tion/lib8tion.h"
-#include "qt/lib8tion/scale8.h"
+#include "lib8tion/lib8tion.h"
+#include "lib8tion/scale8.h"
 
 ///////////////////////////////////////////////////////////////////////
 //
